+++ bigml/tests/test_05_compare_predictions.py
@@ -191,13 +191,16 @@ class TestComparePrediction(object):
                 And I create a proportional missing strategy local prediction for "<data_input>"
                 Then the local prediction is "<prediction>"
                 And the local prediction's confidence is "<confidence>"
+                And the highest local prediction's confidence is "<confidence>"
 
                 Examples:
                 | data               | time_1  | time_2 | time_3 | data_input           | objective | prediction     | confidence |
         """
         examples = [
             ['data/iris_missing2.csv', '10', '10', '10', '{"petal width": 1}', '000004', 'Iris-setosa', '0.8064'],
-            ['data/iris_missing2.csv', '10', '10', '10', '{"petal width": 1, "petal length": 4}', '000004', 'Iris-versicolor', '0.7847']]
+            ['data/iris_missing2.csv', '10', '10', '10', '{"petal width": 1, "petal length": 4}', '000004', 'Iris-versicolor', '0.7847'],
+            ['data/missings_reg.csv', '10', '10', '10', '{"x2": 4}', '000002', '1.33333', '1.62547']
+]
         show_doc(self.test_scenario4, examples)
         for example in examples:
             print "\nTesting with:\n", example
@@ -214,6 +217,7 @@ class TestComparePrediction(object):
             prediction_compare.i_create_a_proportional_local_prediction(self, example[4])
             prediction_compare.the_local_prediction_is(self, example[6])
             prediction_compare.the_local_prediction_confidence_is(self, example[7])
+            prediction_compare.the_highest_local_prediction_confidence_is(self, example[4], example[7])
 
     def test_scenario5(self):
         """
@@ -804,3 +808,45 @@ class TestComparePrediction(object):
             prediction_compare.i_create_a_local_prediction(self, example[4])
             prediction_compare.the_local_prediction_is(self, example[5])
             prediction_compare.the_local_probability_is(self, example[6])
+
+    def test_scenario18(self):
+        """
+            Scenario: Successfully comparing predictions with proportional missing strategy for missing_splits models:
+                Given I create a data source uploading a "<data>" file
+                And I wait until the source is ready less than <time_1> secs
+                And I create a dataset
+                And I wait until the dataset is ready less than <time_2> secs
+                And I create a weighted model with missing splits
+                And I wait until the model is ready less than <time_3> secs
+                And I create a local model
+                When I create a proportional missing strategy prediction for "<data_input>"
+                Then the prediction for "<objective>" is "<prediction>"
+                And the confidence for the prediction is "<confidence>"
+                And I create a proportional missing strategy local prediction for "<data_input>"
+                Then the local prediction is "<prediction>"
+                And the local prediction's confidence is "<confidence>"
+                And the highest local prediction's confidence is "<confidence>"
+
+                Examples:
+                | data               | time_1  | time_2 | time_3 | data_input           | objective | prediction     | confidence |
+        """
+        examples = [
+            ['data/missings_cat.csv', '10', '10', '10', '{"x2": 4}', '000002', 'positive', '0.25241']
+]
+        show_doc(self.test_scenario18, examples)
+        for example in examples:
+            print "\nTesting with:\n", example
+            source_create.i_upload_a_file(self, example[0])
+            source_create.the_source_is_finished(self, example[1])
+            dataset_create.i_create_a_dataset(self)
+            dataset_create.the_dataset_is_finished_in_less_than(self, example[2])
+            model_create.i_create_a_weighted_model_with_missing_splits(self)
+            model_create.the_model_is_finished_in_less_than(self, example[3])
+            prediction_compare.i_create_a_local_model(self)
+            prediction_create.i_create_a_proportional_prediction(self, example[4])
+            prediction_create.the_prediction_is(self, example[5], example[6])
+            prediction_create.the_confidence_is(self, example[7])
+            prediction_compare.i_create_a_proportional_local_prediction(self, example[4])
+            prediction_compare.the_local_prediction_is(self, example[6])
+            prediction_compare.the_local_prediction_confidence_is(self, example[7])
+            prediction_compare.the_highest_local_prediction_confidence_is(self, example[4], example[7])